\name{gen}
\alias{gen}
\title{Genotype data}
\usage{gen}
\description{
	Genotype data of 1003 subjects on 1000 simulated biallelic genetic variants.
}
\format{
	1003x1000 matrix with rows representing genetic markers and columns representing subjects.
}
\details{
	Please refer to the section of simulation study in the paper (Sun et al, BMC genetics, 2014) for details.
}