\name{phe}
\alias{phe}
\title{Phenotype data}
\usage{phe}
\description{
	Phenotype data of 1003 subjects on 10 simulated phenotypic variables.
}
\format{
	1003x10 matrix with rows representing subjects and columns representing phenotypic variables.
}
\details{
	Please refer to the section of simulation study in the paper (Sun et al, BMC genetics, 2014) for details.
}