\name{view2}
\alias{view2}
\title{View2 data}
\usage{view2}
\description{
	The other one of the two views simulated to mimic datasets from a real study, in which genes are characterized with expression patterns.
}
\format{
	999x12 matrix with rows representing genes and columns representing various sources, such as different tissues or time points.
}
\details{
	Please refer to the details included in the description for dataset view1.
}